Comment and clean controllers and clients.

+++ spring-api/src/main/kotlin/com/r3/gallery/broker/corda/client/token/controllers/TokenNetworkBuyerController.kt
@@ -12,6 +12,9 @@ import org.springframework.http.MediaType
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.*
 
+/**
+ * REST endpoints for Buyers on a consideration (GBP or CBDC) network.
+ */
 @CrossOrigin
 @RestController
 @RequestMapping("/buyer")
@@ -22,17 +25,36 @@ class TokenNetworkBuyerController(private val buyerClient: TokenNetworkBuyerClie
         const val TIMEOUT = ConnectionServiceImpl.TIMEOUT
     }
 
+    /**
+     * REST endpoint to Issue tokens on a consideration network
+     *
+     * TODO - NOTE: not currently supported in UI. This endpoint is for future use.
+     *
+     * @param buyerParty to issue the tokens
+     * @param amount of tokens
+     * @param currency description of the token type
+     */
     @PutMapping("/issue-tokens")
     fun issueTokens(
         @RequestParam("buyerParty") buyerParty: TokenParty,
-        @RequestParam("amount") amount: Long,
+        @RequestParam("amount") amount: Int,
         @RequestParam("currency") currency: String
     ): ResponseEntity<Unit> {
         logger.info("Request by $buyerParty to issue $amount $currency to self")
-        buyerClient.issueTokens(buyerParty, amount, currency)
+        buyerClient.issueTokens(buyerParty, amount.toLong(), currency)
         return asResponse(Unit)
     }
 
+    /**
+     * TODO: This is an intermediate request endpoint used to test and will move to BidService placeBid
+     * REST endpoint to transfer encumbered tokens after receiving a draft artwork transfer response
+     *
+     * @param buyerParty to create the encumbered transaction
+     * @param sellerParty who will receive the encumbered tokens
+     * @param amount of tokens
+     * @param currency description of the token type
+     * @param validatedUnsignedArtworkTransferTx to use for the lock on the encumbrance
+     */
     @PostMapping("/transfer-encumbered-tokens", consumes = [MediaType.APPLICATION_JSON_VALUE])
     fun transferEncumberedTokens(
         @RequestParam("buyerParty") buyerParty: TokenParty,
@@ -47,8 +69,14 @@ class TokenNetworkBuyerController(private val buyerClient: TokenNetworkBuyerClie
         return asResponse(signedTokenTransferTxId)
     }
 
-    /*
-     * Releases the unspent encumbered tokens offer specified by encumberedTokensTxHash on the party that issued it
+    /**
+     * TODO: Note, this manual release is not currently implemented in the UI. The scenario is if the seller accepts no winner.
+     * REST endpoint for releasing tokens at buyer request after an auction expiry where the seller did not respond
+     * with a proof-of-action.
+     *
+     * @param buyerParty requesting release
+     * @param currency of tokens
+     * @param encumberedTokensTxHash hash of the encumbered tokens transaction.
      */
     @PostMapping("/release-encumbered-tokens")
     fun transferEncumberedTokens(
